:books: Doxygen - AngelScript doc audit.
Added missing objects: CacheSystemClass, CacheEntryClass.
Added missing enums: SpecialObjectType, FreeForceType
Fixed incomplete  bindings+docs: LoaderType (missing fields ADDONPART and onwards), MsgType (missing FREEFORCE requests).
Monster file "docs/angelscript/Script2Game/globals.h" finally partitioned to individual enums -> one file per enum placed in "enums/".
