+++ doc/angelscript/Script2Game/CVarClass.h
@@ -8,19 +8,6 @@ namespace Script2Game {
 /** \addtogroup Script2Game
  *  @{
  */    
-    
-/**
- * @brief Types and special attributes of cvars.
- * @note Default cvar type is string - To create a string cvar, enter '0' as flags.
- */
-enum CVarFlags
-{
-    CVAR_TYPE_BOOL    = BITMASK(1),
-    CVAR_TYPE_INT     = BITMASK(2),
-    CVAR_TYPE_FLOAT   = BITMASK(3),
-    CVAR_ARCHIVE      = BITMASK(4),    //!< Will be written to RoR.cfg
-    CVAR_NO_LOG       = BITMASK(5)     //!< Will not be written to RoR.log
-};
 
 /**
  * @brief Binding of RoR::CVar; A console variable, usually defined in RoR.cfg but also created by users or scripts.
