+++ doc/angelscript/Script2Game/enums/Ai_events.h
@@ -1,0 +1,30 @@
+
+  // =================================================== //
+  // THIS IS NOT A C++ HEADER! Only a dummy for Doxygen. //
+  // =================================================== //
+
+/** \addtogroup ScriptSideAPIs
+ *  @{
+ */    
+
+/** \addtogroup Script2Game
+ *  @{
+ */   
+
+namespace Script2Game {
+
+/**
+*	AI events, used with Script2Game::VehicleAiClass
+*/
+enum Ai_events
+{
+	AI_HORN,
+	AI_LIGHTSTOGGLE,
+	AI_WAIT_SECONDS,
+	AI_BEACONSTOGGLE
+};
+
+} // namespace Script2Game
+
+/// @}    //addtogroup Script2Game
+/// @}    //addtogroup ScriptSideAPIs
