+++ doc/angelscript/Script2Game/enums/Ai_values.h
@@ -1,0 +1,28 @@
+
+  // =================================================== //
+  // THIS IS NOT A C++ HEADER! Only a dummy for Doxygen. //
+  // =================================================== //
+
+/** \addtogroup ScriptSideAPIs
+ *  @{
+ */    
+
+/** \addtogroup Script2Game
+ *  @{
+ */   
+
+namespace Script2Game {
+
+/**
+*	AI values that can be set, used with Script2Game::VehicleAiClass
+*/
+enum Ai_values
+{
+	AI_SPEED,
+	AI_POWER
+};
+
+} // namespace Script2Game
+
+/// @}    //addtogroup Script2Game
+/// @}    //addtogroup ScriptSideAPIs
