+++ doc/angelscript/Script2Game/enums/BlinkType.h
@@ -1,0 +1,29 @@
+
+  // =================================================== //
+  // THIS IS NOT A C++ HEADER! Only a dummy for Doxygen. //
+  // =================================================== //
+
+/** \addtogroup ScriptSideAPIs
+ *  @{
+ */    
+
+/** \addtogroup Script2Game
+ *  @{
+ */   
+
+namespace Script2Game {
+
+/**
+ * Binding of RoR::BlinkType
+ */
+enum BlinkType {
+    BLINK_NONE,
+    BLINK_LEFT,
+    BLINK_RIGHT,
+    BLINK_WARN,
+}
+
+} // namespace Script2Game
+
+/// @}    //addtogroup Script2Game
+/// @}    //addtogroup ScriptSideAPIs
