+++ doc/angelscript/Script2Game/enums/FlareType.h
@@ -1,0 +1,37 @@
+
+  // =================================================== //
+  // THIS IS NOT A C++ HEADER! Only a dummy for Doxygen. //
+  // =================================================== //
+
+/** \addtogroup ScriptSideAPIs
+ *  @{
+ */    
+
+/** \addtogroup Script2Game
+ *  @{
+ */   
+
+namespace Script2Game {
+
+/**
+ * Binding of RoR::FlareType
+ */
+enum FlareType {
+    FLARE_TYPE_NONE,
+    FLARE_TYPE_HEADLIGHT,
+    FLARE_TYPE_HIGH_BEAM,
+    FLARE_TYPE_FOG_LIGHT,
+    FLARE_TYPE_TAIL_LIGHT,
+    FLARE_TYPE_BRAKE_LIGHT,
+    FLARE_TYPE_REVERSE_LIGHT,
+    FLARE_TYPE_SIDELIGHT,
+    FLARE_TYPE_BLINKER_LEFT,
+    FLARE_TYPE_BLINKER_RIGHT,
+    FLARE_TYPE_USER,
+    FLARE_TYPE_DASHBOARD,
+};
+
+} // namespace Script2Game
+
+/// @}    //addtogroup Script2Game
+/// @}    //addtogroup ScriptSideAPIs
