+++ doc/angelscript/Script2Game/enums/ModulationSources.h
@@ -1,0 +1,56 @@
+
+  // =================================================== //
+  // THIS IS NOT A C++ HEADER! Only a dummy for Doxygen. //
+  // =================================================== //
+
+/** \addtogroup ScriptSideAPIs
+ *  @{
+ */    
+
+/** \addtogroup Script2Game
+ *  @{
+ */   
+
+namespace Script2Game {
+
+/**
+* Binding of `RoR::ModulationSources`
+*/
+enum ModulationSources {
+    SS_MOD_NONE,
+    SS_MOD_ENGINE,
+    SS_MOD_TURBO,
+    SS_MOD_AEROENGINE1,
+    SS_MOD_AEROENGINE2,
+    SS_MOD_AEROENGINE3,
+    SS_MOD_AEROENGINE4,
+    SS_MOD_WHEELSPEED,
+    SS_MOD_INJECTOR,
+    SS_MOD_TORQUE,
+    SS_MOD_GEARBOX,
+    SS_MOD_CREAK,
+    SS_MOD_BREAK,
+    SS_MOD_SCREETCH,
+    SS_MOD_PUMP,
+    SS_MOD_THROTTLE1,
+    SS_MOD_THROTTLE2,
+    SS_MOD_THROTTLE3,
+    SS_MOD_THROTTLE4,
+    SS_MOD_THROTTLE5,
+    SS_MOD_THROTTLE6,
+    SS_MOD_THROTTLE7,
+    SS_MOD_THROTTLE8,
+    SS_MOD_AEROENGINE5,
+    SS_MOD_AEROENGINE6,
+    SS_MOD_AEROENGINE7,
+    SS_MOD_AEROENGINE8,
+    SS_MOD_AIRSPEED,
+    SS_MOD_AOA,
+    SS_MOD_LINKED_COMMANDRATE,
+    SS_MOD_MUSIC_VOLUME
+};
+
+} // namespace Script2Game
+
+/// @}    //addtogroup Script2Game
+/// @}    //addtogroup ScriptSideAPIs
