+++ doc/angelscript/Script2Game/enums/RoadType.h
@@ -1,0 +1,33 @@
+
+  // =================================================== //
+  // THIS IS NOT A C++ HEADER! Only a dummy for Doxygen. //
+  // =================================================== //
+
+/** \addtogroup ScriptSideAPIs
+ *  @{
+ */    
+
+/** \addtogroup Script2Game
+ *  @{
+ */   
+
+namespace Script2Game {
+
+/**
+* Binding of RoR::RoadType, used with Script2Game::ProceduralRoadClass
+*/
+enum RoadType
+{
+    ROAD_AUTOMATIC,
+    ROAD_FLAT,
+    ROAD_LEFT,
+    ROAD_RIGHT,
+    ROAD_BOTH,
+    ROAD_BRIDGE,
+    ROAD_MONORAIL
+};
+
+} // namespace Script2Game
+
+/// @}    //addtogroup Script2Game
+/// @}    //addtogroup ScriptSideAPIs
