+++ doc/angelscript/Script2Game/enums/ScriptRetCode.h
@@ -1,0 +1,65 @@
+
+  // =================================================== //
+  // THIS IS NOT A C++ HEADER! Only a dummy for Doxygen. //
+  // =================================================== //
+
+/** \addtogroup ScriptSideAPIs
+ *  @{
+ */    
+
+/** \addtogroup Script2Game
+ *  @{
+ */   
+
+namespace Script2Game {
+
+/// Binding of `RoR::ScriptRetCode`; Common return codes for script manipulation funcs (add/get/delete | funcs/variables).
+/// Fields named `SCRIPTRETCODE_AS_*` are AngelScript technical codes, the remaining are `RoR::ScriptEngine` return codes.
+enum ScriptRetCode
+{
+    
+    SCRIPTRETCODE_SUCCESS = AngelScript::asSUCCESS, //!< Generic success - 0 by common convention.
+
+    // AngelScript technical codes
+    SCRIPTRETCODE_AS_ERROR = AngelScript::asERROR,
+    SCRIPTRETCODE_AS_CONTEXT_ACTIVE = AngelScript::asCONTEXT_ACTIVE,
+    SCRIPTRETCODE_AS_CONTEXT_NOT_FINISHED = AngelScript::asCONTEXT_NOT_FINISHED,
+    SCRIPTRETCODE_AS_CONTEXT_NOT_PREPARED = AngelScript::asCONTEXT_NOT_PREPARED,
+    SCRIPTRETCODE_AS_INVALID_ARG = AngelScript::asINVALID_ARG,
+    SCRIPTRETCODE_AS_NO_FUNCTION = AngelScript::asNO_FUNCTION,
+    SCRIPTRETCODE_AS_NOT_SUPPORTED = AngelScript::asNOT_SUPPORTED,
+    SCRIPTRETCODE_AS_INVALID_NAME = AngelScript::asINVALID_NAME,
+    SCRIPTRETCODE_AS_NAME_TAKEN = AngelScript::asNAME_TAKEN,
+    SCRIPTRETCODE_AS_INVALID_DECLARATION = AngelScript::asINVALID_DECLARATION,
+    SCRIPTRETCODE_AS_INVALID_OBJECT = AngelScript::asINVALID_OBJECT,
+    SCRIPTRETCODE_AS_INVALID_TYPE = AngelScript::asINVALID_TYPE,
+    SCRIPTRETCODE_AS_ALREADY_REGISTERED = AngelScript::asALREADY_REGISTERED,
+    SCRIPTRETCODE_AS_MULTIPLE_FUNCTIONS = AngelScript::asMULTIPLE_FUNCTIONS,
+    SCRIPTRETCODE_AS_NO_MODULE = AngelScript::asNO_MODULE,
+    SCRIPTRETCODE_AS_NO_GLOBAL_VAR = AngelScript::asNO_GLOBAL_VAR,
+    SCRIPTRETCODE_AS_INVALID_CONFIGURATION = AngelScript::asINVALID_CONFIGURATION,
+    SCRIPTRETCODE_AS_INVALID_INTERFACE = AngelScript::asINVALID_INTERFACE,
+    SCRIPTRETCODE_AS_CANT_BIND_ALL_FUNCTIONS = AngelScript::asCANT_BIND_ALL_FUNCTIONS,
+    SCRIPTRETCODE_AS_LOWER_ARRAY_DIMENSION_NOT_REGISTERED = AngelScript::asLOWER_ARRAY_DIMENSION_NOT_REGISTERED,
+    SCRIPTRETCODE_AS_WRONG_CONFIG_GROUP = AngelScript::asWRONG_CONFIG_GROUP,
+    SCRIPTRETCODE_AS_CONFIG_GROUP_IS_IN_USE = AngelScript::asCONFIG_GROUP_IS_IN_USE,
+    SCRIPTRETCODE_AS_ILLEGAL_BEHAVIOUR_FOR_TYPE = AngelScript::asILLEGAL_BEHAVIOUR_FOR_TYPE,
+    SCRIPTRETCODE_AS_WRONG_CALLING_CONV = AngelScript::asWRONG_CALLING_CONV,
+    SCRIPTRETCODE_AS_BUILD_IN_PROGRESS = AngelScript::asBUILD_IN_PROGRESS,
+    SCRIPTRETCODE_AS_INIT_GLOBAL_VARS_FAILED = AngelScript::asINIT_GLOBAL_VARS_FAILED,
+    SCRIPTRETCODE_AS_OUT_OF_MEMORY = AngelScript::asOUT_OF_MEMORY,
+    SCRIPTRETCODE_AS_MODULE_IS_IN_USE = AngelScript::asMODULE_IS_IN_USE,
+
+    // RoR ScriptEngine return codes
+    SCRIPTRETCODE_UNSPECIFIED_ERROR = -1001,
+    SCRIPTRETCODE_ENGINE_NOT_CREATED = -1002,
+    SCRIPTRETCODE_CONTEXT_NOT_CREATED = -1003,
+    SCRIPTRETCODE_SCRIPTUNIT_NOT_EXISTS = -1004,
+    SCRIPTRETCODE_SCRIPTUNIT_NO_MODULE = -1005,
+    SCRIPTRETCODE_FUNCTION_NOT_EXISTS = -1006,
+};
+
+} // namespace Script2Game
+
+/// @}    //addtogroup Script2Game
+/// @}    //addtogroup ScriptSideAPIs
