+++ doc/angelscript/Script2Game/enums/TextureFit.h
@@ -1,0 +1,37 @@
+
+  // =================================================== //
+  // THIS IS NOT A C++ HEADER! Only a dummy for Doxygen. //
+  // =================================================== //
+
+/** \addtogroup ScriptSideAPIs
+ *  @{
+ */    
+
+/** \addtogroup Script2Game
+ *  @{
+ */   
+
+namespace Script2Game {
+
+/**
+* Binding of RoR::TextureFit, used with Script2Game::ProceduralRoadClass
+*/
+enum TextureFit
+{
+    TEXFIT_NONE,
+    TEXFIT_BRICKWALL,
+    TEXFIT_ROADS1,
+    TEXFIT_ROADS2,
+    TEXFIT_ROAD,
+    TEXFIT_ROADS3,
+    TEXFIT_ROADS4,
+    TEXFIT_CONCRETEWALL,
+    TEXFIT_CONCRETEWALLI,
+    TEXFIT_CONCRETETOP,
+    TEXFIT_CONCRETEUNDER
+};
+
+} // namespace Script2Game
+
+/// @}    //addtogroup Script2Game
+/// @}    //addtogroup ScriptSideAPIs
