+++ doc/angelscript/Script2Game/enums/autoswitch.h
@@ -1,0 +1,32 @@
+
+  // =================================================== //
+  // THIS IS NOT A C++ HEADER! Only a dummy for Doxygen. //
+  // =================================================== //
+
+/** \addtogroup ScriptSideAPIs
+ *  @{
+ */    
+
+/** \addtogroup Script2Game
+ *  @{
+ */   
+
+namespace Script2Game {
+
+/**
+* Binding of RoR::autoswitch, used with Script2Game::EngineClass
+*/
+enum autoswitch
+{
+    REAR,
+    NEUTRAL,
+    DRIVE,
+    TWO,
+    ONE,
+    MANUALMODE
+};
+
+} // namespace Script2Game
+
+/// @}    //addtogroup Script2Game
+/// @}    //addtogroup ScriptSideAPIs
